Kfold changes

+++ Kfold.py
@@ -1,10 +1,6 @@
 import os
 import pandas as pd
-import torch
-import xlwt as xlwt
-from openpyxl.reader.excel import load_workbook
 from tqdm import tqdm
-
 import bert_classifier_train
 from evaluation import evaluate_output, evaluate_classifier, compare_output
 
@@ -14,10 +10,7 @@ fold_path = "Kfolds"
 folds = os.listdir(fold_path)
 folds.sort()
 
-# book = xlwt.Workbook(encoding="utf-8")
-#
-# sheet = book.add_sheet(bert)
-
+best_epoch_list = []
 best_recall_list = []
 best_precision_list = []
 best_threshold_list = []
@@ -45,13 +38,20 @@ for fold in tqdm(folds):
     version_list.append(version)
 
     best_epoch = valid_result.index(min(valid_result)) + 1
+    best_epoch_list.append(best_epoch)
     print(f"Best epoch: {best_epoch}")
 
-    data_path = os.path.join(fold_path, fold, "cns_fulltest_raw.csv")
+    data_path = os.path.join(fold_path, fold, "cns_test_raw.csv")
 
     batch_size = 2  # debug
 
     evaluate_classifier.test(bert, version, best_epoch, data_path, batch_size)
+
+    test_output = pd.read_csv(os.path.join("output", f"{bert}_{version}_epoch{best_epoch}_TEST.csv"))
+    val_output = pd.read_csv(os.path.join("output", f"{bert}_{version}_epoch{best_epoch}_VAL.csv"))
+
+    full_output = test_output.append(val_output)
+    full_output.to_csv(os.path.join("output", f"{bert}_{version}_epoch{best_epoch}.csv"))
 
     output_path = "output"
     precision_list, recall_list, threshold_list = evaluate_output.evaluate(bert, version, best_epoch, output_path)
@@ -59,15 +59,19 @@ for fold in tqdm(folds):
     best_threshold = 0.5
     best_precision = 0
     best_recall = 0
+    min_precision = 0.4
 
-    for i, precision in enumerate(precision_list):
-        recall = recall_list[i]
-        threshold = threshold_list[i]
-        if precision > 0.4 and precision >= best_precision:
-            if recall >= best_recall:
-                best_recall = recall
-                best_precision = precision
-                best_threshold = threshold
+    while best_precision == 0 and min_precision > 0.2:
+        for i, precision in enumerate(precision_list):
+            precision = float(precision)
+            recall = float(recall_list[i])
+            threshold = float(threshold_list[i])
+            if precision > min_precision and precision >= best_precision:
+                if recall >= best_recall:
+                    best_recall = recall
+                    best_precision = precision
+                    best_threshold = threshold
+        min_precision -= 0.05
 
     print(f"best_recall: {best_recall} best_precision: {best_precision} best_threshold: {best_threshold}")
 
@@ -75,7 +79,7 @@ for fold in tqdm(folds):
     best_recall_list.append(best_recall)
     best_precision_list.append(best_precision)
 
-    true_pos, true_neg, false_pos, false_neg = compare_output.compare(best_threshold, bert, version, best_epoch, output_path, fold)
+    true_pos, true_neg, false_pos, false_neg = compare_output.compare(0.5, bert, version, best_epoch, output_path)
 
     fp_list.append(false_pos)
     fn_list.append(false_neg)
@@ -87,6 +91,9 @@ for fold in tqdm(folds):
 Kfold_results = pd.DataFrame({"Fold": folds, "Version": version_list, "Recall": best_recall_list, "Precision": best_precision_list,
                               "Threshold": best_threshold_list, "False Neg": fn_list, "False Pos": fp_list, "Val loss": valid_result_list})
 
-with pd.ExcelWriter("path_to_file.xlsx", if_sheet_exists='overlay') as writer:
-    Kfold_results.to_excel(writer, sheet_name=bert)
 
+if os.path.exists("Kfolds/Kfold_results.xlsx"):
+    with pd.ExcelWriter("path_to_file.xlsx", mode='a', if_sheet_exists='overlay') as writer:
+        Kfold_results.to_excel(writer, sheet_name=bert)
+else:
+    Kfold_results.to_excel("Kfolds/Kfold_results.xlsx", sheet_name=bert)
